Push /pj errors to UI over websocket
Sending errors over websocket to the UI shows them as they happen.
This has a different security considerations from typical error flow
because both the sender and receiver are the same party in nolooking.
Because they're the same party and the app is hosted behind an umbrel
firewall, we use the unsecure ws:// protocol. It's behind Tor anyway.